+++ src/main/java/com/project/professor/allocation/service/ProfessorService.java
@@ -5,25 +5,34 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.project.professor.allocation.entity.Allocation;
+import com.project.professor.allocation.entity.Department;
 import com.project.professor.allocation.entity.Professor;
+import com.project.professor.allocation.repository.AllocationRepository;
 import com.project.professor.allocation.repository.ProfessorRepository;
 
 @Service
 public class ProfessorService {
 
 	private final ProfessorRepository professorRepository;
+	private final AllocationRepository allocationRepository;
+	private final DepartmentService departmentService;
 
-	public ProfessorService(ProfessorRepository professorRepository) {
+	public ProfessorService(ProfessorRepository professorRepository, AllocationRepository allocationRepository, DepartmentService departmentService) {
 		super();
 		this.professorRepository = professorRepository;
+		this.allocationRepository = allocationRepository;
+		this.departmentService = departmentService;
 	}
 
 	// CRUD Ler todos
-	public List<Professor> findAll() {
-
-		List<Professor> professors = professorRepository.findAll();
-		return professors;
-
+	public List<Professor> findAll(String name) {
+		
+		if (name == null) {
+			return professorRepository.findAll();
+		} else {
+			return professorRepository.findByNameContainingIgnoreCase(name);
+		}
 	}
 
 	// CRUD Ler Id
@@ -32,6 +41,12 @@ public class ProfessorService {
 		Optional<Professor> optional = professorRepository.findById(id);
 		Professor professors = optional.orElse(null);
 		return professors;
+
+	}
+	
+	public List <Professor> findByDepartmentId(Long id) {
+
+		return professorRepository.findByDepartmentId(id);
 
 	}
 
@@ -56,16 +71,21 @@ public class ProfessorService {
 	}
 	
 	private Professor saveInternal(Professor professor) {
+        professor = professorRepository.save(professor);
 
-		Professor profNew = professorRepository.save(professor);
-		return profNew;
+        Department department = departmentService.findById(professor.getDepartmentId());
+        professor.setDepart(department);
 
-	}
+        List<Allocation> allocations = allocationRepository.findByProfessorId(professor.getId());
+        professor.setAllocations(allocations);
+
+        return professor;
+    }
 
 	// CRUD Deletar pelo Id
 	public void deleteById(Long id) {
 
-		if (professorRepository.existsById(id)) {
+		if (id != null && professorRepository.existsById(id)) {
 			professorRepository.deleteById(id);
 		}
 	}
